+++ HISTORY.md
@@ -1,5 +1,12 @@
 ## History
 
+### 0.1.x (2023-10-04)
+- Change the API endpoint - as the old one was discontinued
+- Add API access_key configurationn option
+- Change parsed fields of the JSON response: 
+  - "rates" -> "quotes"
+  - "base" -> "source"
+
 ### 0.1.2 (2023-03-29)
 
 - Fix pypi description